requirements.txt
Flask>=2.3.0
pydicom>=2.4.0
Pillow>=10.0.0
numpy>=1.21.0
Werkzeug>=2.3.0